Agrego servicio completo Get and Post de los registros del sensor de calidad del aire

+++ src/main/java/aq/airquality/backend/entidades/Usuarios.java
@@ -2,7 +2,6 @@ package aq.airquality.backend.entidades;
 
 import java.io.Serializable;
 
-import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -17,7 +16,7 @@ public class Usuarios implements Serializable {
 
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private Long id;
+	private int id;
 
 	private String correo;
 	private int altura;
@@ -26,11 +25,11 @@ public class Usuarios implements Serializable {
 	private int cantidaddepasos;
 	private int horasdesueño;
 
-	public Long getId() {
+	public int getId() {
 		return id;
 	}
 
-	public void setId(Long id) {
+	public void setId(int id) {
 		this.id = id;
 	}
